+++ CHANGELOG.md
@@ -1,2 +1,6 @@
 # Changelog
 
+## Breaking Changes
+
+- context.$smtp: Smpt service is removed and replaced with $email.
+- context.\$email: Email service is added with a send() function.